+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/engine/EngineObserver.kt
@@ -13,6 +13,7 @@ import mozilla.components.browser.state.action.CrashAction
 import mozilla.components.browser.state.action.EngineAction
 import mozilla.components.browser.state.action.MediaSessionAction
 import mozilla.components.browser.state.action.ReaderAction
+import mozilla.components.browser.state.action.ShoppingProductAction
 import mozilla.components.browser.state.action.TrackingProtectionAction
 import mozilla.components.browser.state.selector.findTabOrCustomTab
 import mozilla.components.browser.state.state.AppIntentState
@@ -162,6 +163,10 @@ internal class EngineObserver(
 
     override fun onCookieBannerChange(status: EngineSession.CookieBannerHandlingStatus) {
         store.dispatch(CookieBannerAction.UpdateStatusAction(tabId, status))
+    }
+
+    override fun onProductUrlChange(isProductUrl: Boolean) {
+        store.dispatch(ShoppingProductAction.UpdateProductUrlStatusAction(tabId, isProductUrl))
     }
 
     override fun onLongPress(hitResult: HitResult) {