+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/state/SessionState.kt
@@ -24,6 +24,7 @@ import mozilla.components.support.utils.SafeIntent
  * contextual identity to use for the session's cookie store.
  * https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/Work_with_contextual_identities
  * @property restored Indicates if this session was restored from a hydrated state.
+ * @property isProductUrl Indicates if this session is currently displaying a product URL.
  */
 interface SessionState {
     val id: String
@@ -36,6 +37,7 @@ interface SessionState {
     val contextId: String?
     val source: Source
     val restored: Boolean
+    val isProductUrl: Boolean
 
     /**
      * Copy the class and override some parameters.
@@ -50,6 +52,7 @@ interface SessionState {
         mediaSessionState: MediaSessionState? = this.mediaSessionState,
         contextId: String? = this.contextId,
         cookieBanner: CookieBannerHandlingStatus = this.cookieBanner,
+        isProductUrl: Boolean = this.isProductUrl,
     ): SessionState
 
     /**