+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/state/TabSessionState.kt
@@ -31,6 +31,7 @@ import java.util.UUID
  * @property lastMediaAccessState - [LastMediaAccessState] detailing the tab state when media started playing.
  * Requires [LastMediaAccessMiddleware] to update the value when playback starts.
  * @property restored Indicates if this page was restored from a persisted state.
+ * @property isProductUrl has the product URL status of this tab.
  */
 data class TabSessionState(
     override val id: String = UUID.randomUUID().toString(),
@@ -43,6 +44,7 @@ data class TabSessionState(
     override val contextId: String? = null,
     override val source: SessionState.Source = SessionState.Source.Internal.None,
     override val restored: Boolean = false,
+    override val isProductUrl: Boolean = false,
     val parentId: String? = null,
     val lastAccess: Long = 0L,
     val createdAt: Long = System.currentTimeMillis(),
@@ -60,6 +62,7 @@ data class TabSessionState(
         mediaSessionState: MediaSessionState?,
         contextId: String?,
         cookieBanner: CookieBannerHandlingStatus,
+        isProductUrl: Boolean,
     ): SessionState = copy(
         id = id,
         content = content,
@@ -69,6 +72,7 @@ data class TabSessionState(
         mediaSessionState = mediaSessionState,
         contextId = contextId,
         cookieBanner = cookieBanner,
+        isProductUrl = isProductUrl,
     )
 }